Assignment2/requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
bidict==0.22.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.4
cycler==0.11.0
debugpy==1.4.1
decorator==5.0.9
entrypoints==0.3
Flask==2.2.2
Flask-SocketIO==5.3.1
fonttools==4.37.1
frozenlist==1.3.1
idna==3.4
importlib-metadata==4.12.0
ipykernel==6.3.1
ipython==7.27.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.1.2
jupyter-client==7.0.2
jupyter-core==4.7.1
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.3
multidict==6.0.2
nest-asyncio==1.5.1
numpy==1.23.3
packaging==21.3
parso==0.8.2
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.20
Pygments==2.10.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-engineio==4.3.4
python-socketio==5.7.1
pywin32==302
requests==2.28.1
six==1.16.0
tornado==6.1
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
wincertstore==0.2
yarl==1.8.1
zipp==3.8.1
